utils/requirements.txt
# Requirements for LogSplitter Web Dashboard and InfluxDB Collector
paho-mqtt>=1.6.0
influxdb-client>=1.36.0
flask>=2.3.0
flask-cors>=4.0.0
flask-socketio>=5.3.0
eventlet>=0.33.0
flask>=2.3.0
flask-socketio>=5.3.0
plotly>=5.15.0
pandas>=1.5.0
argparse